chore: Inline runs routes into the main router file
