+++ control-plane/src/modules/router.ts
@@ -4,7 +4,6 @@ import fs from "fs";
 import path from "path";
 import { ulid } from "ulid";
 import util from "util";
-import { deleteAgent, getAgent, listAgents, upsertAgent, validateSchema } from "./agents";
 import { getBlobData } from "./blobs";
 import * as data from "./data";
 import { integrationsRouter } from "./integrations/router";
@@ -14,7 +13,6 @@ import * as events from "./observability/events";
 import { posthog } from "./posthog";
 import { addMessageAndResume } from "./runs";
 import { getRunMessagesForDisplayWithPolling } from "./runs/messages";
-import { runsRouter } from "./runs/router";
 import { getServiceDefinitions } from "./service-definitions";
 import { unqualifiedEntityId } from "./auth/auth";
 import { upsertMachine } from "./machines";
@@ -32,6 +30,23 @@ import { createBlob } from "./blobs";
 import { recordServicePoll } from "./service-definitions";
 import { getJob } from "./jobs/jobs";
 import { BadRequestError, NotFoundError } from "../utilities/errors";
+import {
+  createRun,
+  deleteRun,
+  getClusterRuns,
+  getAgentMetrics,
+  getRunDetails,
+  addMessageAndResumeWithRun,
+  updateRunFeedback,
+} from "./runs";
+import { RunOptions, mergeAgentOptions, deleteAgent, getAgent, listAgents, upsertAgent, validateSchema } from "./agents";
+import { normalizeFunctionReference } from "./service-definitions";
+import { getClusterDetails } from "./cluster";
+import { JsonSchemaInput } from "inferable/bin/types";
+import { getRunsByTag } from "./runs/tags";
+import { timeline } from "./timeline";
+import { getBlobsForJobs } from "./blobs";
+import { getJobReferences } from "./jobs/jobs";
 
 const readFile = util.promisify(fs.readFile);
 
@@ -109,7 +124,415 @@ export const router = initServer().router(contract, {
       },
     };
   },
-  ...runsRouter.routes,
+  getRun: async request => {
+    const { clusterId, runId } = request.params;
+    const auth = request.request.getAuth();
+    await auth.canAccess({ run: { clusterId, runId } });
+
+    const run = await getRunDetails({
+      clusterId,
+      runId,
+    });
+
+    if (!run.id) {
+      return {
+        status: 404,
+      };
+    }
+
+    return {
+      status: 200,
+      body: {
+        ...run,
+        result: run.result ?? null,
+      },
+    };
+  },
+  createRun: async request => {
+    const { clusterId } = request.params;
+    const body = request.body;
+
+    const auth = request.request.getAuth();
+    await auth.canAccess({ cluster: { clusterId } });
+    auth.canCreate({ run: true });
+
+    if (body.attachedFunctions && body.attachedFunctions.length == 0) {
+      return {
+        status: 400,
+        body: {
+          message: "attachedFunctions cannot be an empty array",
+        },
+      };
+    }
+
+    if (body.resultSchema) {
+      const validationError = validateSchema({
+        schema: body.resultSchema,
+        name: "resultSchema",
+      });
+      if (validationError) {
+        return validationError;
+      }
+    }
+
+    let onStatusChangeHandler: string | undefined = undefined;
+    let onStatusChangeStatuses: string[] | undefined = undefined;
+
+    if (body.onStatusChange) {
+      if (body.onStatusChange.function && body.onStatusChange.webhook) {
+        return {
+          status: 400,
+          body: {
+            message: "onStatusChange cannot have both function and webhook",
+          },
+        };
+      }
+
+      if (body.onStatusChange.function) {
+        // TODO: Validate that onStatusChange and attachedFunctions exist
+        // TODO: Validate that onStatusChange schema is correct
+        onStatusChangeHandler =
+          body.onStatusChange.function && normalizeFunctionReference(body.onStatusChange.function);
+      }
+
+      if (body.onStatusChange.webhook) {
+        // Check for valid Schema
+        if (!body.onStatusChange.webhook.startsWith("https://")) {
+          return {
+            status: 400,
+            body: {
+              message: "onStatusChange webhook must start with https://",
+            },
+          };
+        }
+
+        onStatusChangeHandler = body.onStatusChange.webhook;
+        onStatusChangeStatuses = body.onStatusChange.statuses ?? ["done", "failed"];
+      }
+    }
+
+    let runOptions: RunOptions & { runId?: string } = {
+      runId: body.runId,
+      initialPrompt: body.initialPrompt,
+      systemPrompt: body.systemPrompt,
+      attachedFunctions: body.attachedFunctions?.map(normalizeFunctionReference),
+      resultSchema: body.resultSchema
+        ? (dereferenceSync(body.resultSchema) as JsonSchemaInput)
+        : undefined,
+      interactive: body.interactive,
+      modelIdentifier: body.model,
+      callSummarization: body.callSummarization,
+      reasoningTraces: body.reasoningTraces,
+      enableResultGrounding: body.enableResultGrounding,
+
+      input: body.input,
+    };
+
+    const agent = body.agentId
+      ? await getAgent({
+          clusterId,
+          id: body.agentId,
+        })
+      : undefined;
+
+    if (body.agentId) {
+      if (!agent) {
+        throw new NotFoundError("Agent not found");
+      }
+
+      const merged = mergeAgentOptions(runOptions, agent);
+
+      if (merged.error) {
+        return merged.error;
+      }
+
+      runOptions = merged.options;
+
+      runOptions.messageMetadata = {
+        displayable: {
+          templateName: agent.name,
+          templateId: agent.id,
+          ...body.input,
+        },
+      };
+    }
+
+    if (runOptions.input) {
+      runOptions.initialPrompt = `${runOptions.initialPrompt}\n\n<DATA>\n${JSON.stringify(runOptions.input, null, 2)}\n</DATA>`;
+    }
+
+    const customAuth = auth.type === "custom" ? auth.isCustomAuth() : undefined;
+
+    const run = await createRun({
+      runId: runOptions.runId,
+      userId: auth.entityId,
+      clusterId,
+
+      name: body.name,
+      test: body.test?.enabled ?? false,
+      testMocks: body.test?.mocks,
+      tags: body.tags,
+
+      agentId: agent?.id,
+
+      // Customer Auth context (In the future all auth types should inject context into the run)
+      authContext: customAuth?.context,
+
+      context: body.context,
+
+      onStatusChangeHandler: onStatusChangeHandler,
+      onStatusChangeStatuses: onStatusChangeStatuses,
+
+      // Merged Options
+      resultSchema: runOptions.resultSchema,
+      enableSummarization: runOptions.callSummarization,
+      modelIdentifier: runOptions.modelIdentifier,
+      interactive: runOptions.interactive,
+      systemPrompt: runOptions.systemPrompt,
+      attachedFunctions: runOptions.attachedFunctions,
+      reasoningTraces: runOptions.reasoningTraces,
+      enableResultGrounding: runOptions.enableResultGrounding,
+    });
+
+    if (runOptions.initialPrompt) {
+      await addMessageAndResumeWithRun({
+        id: ulid(),
+        userId: auth.entityId,
+        clusterId,
+        run,
+        message: runOptions.initialPrompt,
+        type: body.agentId ? "template" : "human",
+        metadata: runOptions.messageMetadata,
+        skipAssert: true,
+      });
+    }
+
+    const cluster = await getClusterDetails(clusterId);
+
+    posthog?.capture({
+      distinctId: unqualifiedEntityId(auth.entityId),
+      event: "api:run_create",
+      groups: {
+        organization: auth.organizationId,
+        cluster: clusterId,
+      },
+      properties: {
+        cluster_id: clusterId,
+        is_demo: cluster.is_demo,
+        run_id: run.id,
+        agent_id: run.agentId,
+        cli_version: request.headers["x-cli-version"],
+        user_agent: request.headers["user-agent"],
+      },
+    });
+
+    return {
+      status: 201,
+      body: { id: run.id },
+    };
+  },
+  deleteRun: async request => {
+    const { clusterId, runId } = request.params;
+
+    const auth = request.request.getAuth();
+    await auth.canManage({ run: { clusterId, runId } });
+
+    await deleteRun({
+      clusterId,
+      runId,
+    });
+
+    posthog?.capture({
+      distinctId: unqualifiedEntityId(auth.entityId),
+      event: "api:run_delete",
+      groups: {
+        organization: auth.organizationId,
+        cluster: clusterId,
+      },
+      properties: {
+        cluster_id: clusterId,
+        run_id: runId,
+        cli_version: request.headers["x-cli-version"],
+        user_agent: request.headers["user-agent"],
+      },
+    });
+
+    return {
+      status: 204,
+      body: undefined,
+    };
+  },
+  createFeedback: async request => {
+    const { clusterId, runId } = request.params;
+    const { comment, score } = request.body;
+
+    const auth = request.request.getAuth();
+    await auth.canManage({ run: { clusterId, runId } });
+
+    await updateRunFeedback({
+      id: runId,
+      clusterId,
+      feedbackComment: comment ?? undefined,
+      feedbackScore: score ?? undefined,
+    });
+
+    events.write({
+      type: "runFeedbackSubmitted",
+      clusterId,
+      workflowId: runId,
+      userId: auth.entityId,
+      meta: {
+        feedbackScore: score ?? undefined,
+        feedbackComment: comment ?? undefined,
+      },
+    });
+
+    posthog?.capture({
+      distinctId: unqualifiedEntityId(auth.entityId),
+      event: "api:feedback_create",
+      groups: {
+        organization: auth.organizationId,
+        cluster: clusterId,
+      },
+      properties: {
+        cluster_id: clusterId,
+        run_id: runId,
+        score: score,
+        cli_version: request.headers["x-cli-version"],
+        user_agent: request.headers["user-agent"],
+      },
+    });
+
+    return {
+      status: 204,
+      body: undefined,
+    };
+  },
+  listRuns: async request => {
+    const { clusterId } = request.params;
+    const { test, limit, tags, agentId } = request.query;
+    let { userId } = request.query;
+
+    const auth = request.request.getAuth();
+    await auth.canAccess({ cluster: { clusterId } });
+
+    // Custom auth can only access their own Runs
+    if (auth.type === "custom") {
+      userId = auth.entityId;
+    }
+
+    if (tags) {
+      // ?meta=key:value
+      const [key, value] = tags.split(":");
+
+      if (!key || !value) {
+        return {
+          status: 400,
+          body: {
+            message: "Invalid tag filter format",
+          },
+        };
+      }
+
+      const result = await getRunsByTag({
+        clusterId,
+        key,
+        value,
+        limit,
+        agentId,
+        userId,
+      });
+
+      return {
+        status: 200,
+        body: result,
+      };
+    }
+
+    const result = await getClusterRuns({
+      clusterId,
+      userId,
+      test: test ?? false,
+      limit,
+      agentId,
+    });
+
+    return {
+      status: 200,
+      body: result,
+    };
+  },
+  getRunTimeline: async request => {
+    const { clusterId, runId } = request.params;
+    const { messagesAfter, activityAfter } = request.query;
+
+    const auth = request.request.getAuth();
+    await auth.canAccess({ run: { clusterId, runId } });
+
+    const { messages, activity, jobs, run } = await timeline.getRunTimeline({
+      clusterId,
+      runId,
+      messagesAfter,
+      activityAfter,
+    });
+
+    if (!run) {
+      return {
+        status: 404,
+      };
+    }
+
+    const blobs = await getBlobsForJobs({
+      clusterId,
+      jobIds: jobs.map(job => job.id),
+    });
+
+    return {
+      status: 200,
+      body: {
+        messages,
+        activity,
+        jobs,
+        run,
+        blobs,
+      },
+    };
+  },
+  getAgentMetrics: async request => {
+    const { clusterId, agentId } = request.params;
+
+    const auth = request.request.getAuth();
+    await auth.canAccess({ cluster: { clusterId } });
+
+    const result = await getAgentMetrics({
+      clusterId,
+      agentId,
+    });
+
+    return {
+      status: 200,
+      body: result,
+    };
+  },
+  listRunReferences: async request => {
+    const { clusterId, runId } = request.params;
+    const { token, before } = request.query;
+
+    const auth = request.request.getAuth();
+    await auth.canAccess({ run: { clusterId, runId } });
+
+    const jobReferences = await getJobReferences({
+      clusterId,
+      runId,
+      token,
+      before: new Date(before),
+    });
+
+    return {
+      status: 200,
+      body: jobReferences,
+    };
+  },
   createApiKey: async (request) => {
     const { name } = request.body;
     const { clusterId } = request.params;
